requirements.txt
openai>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic-ai-slim[openai]>=1.87.0
pydantic-settings>=2.12.0
httpx>=0.28.0
trafilatura>=2.0.0
sse-starlette>=2.4.1,<3.0.0
pluggy>=1.6.0
serpapi>=1.0.0
